Create slide-animation-min.js
Create a minified version of the script for on-scroll animations.
The script adds a "slide-animation" class to web elements to prepare them for animation.
An "IntersectionObserver" is implemented to monitor selected elements. When an element becomes at least 25% visible in the viewport, the "visible" class is added to trigger the animation.