Update elements-min.js
Update the minified version of the library of custom reusable web components for Move It.
Enhance the library containing my collection of custom, reusable web components that are used across different parts of the project. These components go beyond what browsers provide by default, allowing for expanded capabilities and functionality in the project.